polish(sidebar): apply code-review findings to pasteboard codec
Splits the AppKit pasteboard extension out of `SidebarView+Groups.swift`
into a dedicated `DraggableSidebarItem+Pasteboard.swift` so each file
holds a single primary concern, and converts the parameterless
`pasteboardType()` helper into a `pasteboardType` static property to
match Apple's API design guidance. Also rewrites the test suite's
doc-comment to describe the present invariant rather than the release
milestone.
